[repo] Factor out parser and error code into their own packages
