+++ samlang-core-errors/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "samlang-core-errors",
+  "version": "0.0.1",
+  "license": "AGPLv3",
+  "scripts": {
+    "compile": "tsc --incremental --noEmit"
+  },
+  "dependencies": {
+    "samlang-core-ast": "workspace:0.0.1"
+  },
+  "devDependencies": {
+    "@types/jest": "^26.0.10",
+    "typescript": "4.0.2"
+  }
+}
