+++ samlang-core-parser/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "samlang-core-parser",
+  "version": "0.0.1",
+  "license": "AGPLv3",
+  "scripts": {
+    "compile": "tsc --incremental --noEmit",
+    "codegen": "yarn antlr4ts -o generated -visitor -no-listener PL.g4 && node suppress-type-errors.js"
+  },
+  "dependencies": {
+    "antlr4ts": "0.5.0-alpha.3",
+    "samlang-core-ast": "workspace:0.0.1",
+    "samlang-core-errors": "workspace:0.0.1",
+    "samlang-core-utils": "workspace:0.0.1"
+  },
+  "devDependencies": {
+    "@types/jest": "^26.0.10",
+    "antlr4ts-cli": "0.5.0-alpha.3",
+    "typescript": "4.0.2"
+  }
+}
