+++ samlang-core-parser/parser-module-builder.ts
@@ -1,6 +1,5 @@
 import { AbstractParseTreeVisitor } from 'antlr4ts/tree/AbstractParseTreeVisitor';
 
-import type { ModuleErrorCollector } from '../errors';
 import type {
   ModuleContext,
   ImportModuleMembersContext,
@@ -17,6 +16,7 @@ import type {
   ModuleMembersImport,
   SamlangModule,
 } from 'samlang-core-ast/samlang-toplevel';
+import type { ModuleErrorCollector } from 'samlang-core-errors';
 import { isNotNull, assertNotNull } from 'samlang-core-utils';
 
 class ModuleMemberVisitor
