+++ samlang-core/checker/__tests/constraint-aware-checker.test.ts
@@ -1,4 +1,3 @@
-import { createGlobalErrorCollector } from '../../errors';
 import { checkAndInfer, ConstraintAwareChecker } from '../constraint-aware-checker';
 import TypeResolution from '../type-resolution';
 
@@ -13,6 +12,7 @@ import {
   Range,
   ModuleReference,
 } from 'samlang-core-ast/common-nodes';
+import { createGlobalErrorCollector } from 'samlang-core-errors';
 
 it('t1=primitive type', () => {
   const resolution = new TypeResolution();
