+++ samlang-core/checker/__tests/undefined-import-checker.test.ts
@@ -1,8 +1,8 @@
-import { createGlobalErrorCollector } from '../../errors';
 import checkUndefinedImportsError from '../undefined-imports-checker';
 
 import { ModuleReference, Range, Sources } from 'samlang-core-ast/common-nodes';
 import type { SamlangModule, ClassDefinition } from 'samlang-core-ast/samlang-toplevel';
+import { createGlobalErrorCollector } from 'samlang-core-errors';
 import { hashMapOf } from 'samlang-core-utils';
 
 const createMockClass = (name: string, isPublic: boolean): ClassDefinition => ({
