+++ samlang-core/checker/statement-type-checker.ts
@@ -1,4 +1,3 @@
-import type { ModuleErrorCollector } from '../errors';
 import type { AccessibleGlobalTypingContext, LocalTypingContext } from './typing-context';
 
 import { Type, unitType } from 'samlang-core-ast/common-nodes';
@@ -11,6 +10,7 @@ import {
 } from 'samlang-core-ast/samlang-expressions';
 import type { Pattern, ObjectPatternDestucturedName } from 'samlang-core-ast/samlang-pattern';
 import type { FieldType } from 'samlang-core-ast/samlang-toplevel';
+import type { ModuleErrorCollector } from 'samlang-core-errors';
 import { isNotNull, assertNotNull } from 'samlang-core-utils';
 
 export default class StatementTypeChecker {
