+++ samlang-core/checker/undefined-imports-checker.ts
@@ -1,7 +1,6 @@
-import type { ModuleErrorCollector } from '../errors';
-
 import type { Sources } from 'samlang-core-ast/common-nodes';
 import type { ModuleMembersImport, SamlangModule } from 'samlang-core-ast/samlang-toplevel';
+import type { ModuleErrorCollector } from 'samlang-core-errors';
 import { isNotNull } from 'samlang-core-utils';
 
 class UndefinedImportChecker {
