+++ samlang-core/index.ts
@@ -1,5 +1,4 @@
 export { compileSamlangSourcesToHighIRSources } from './compiler';
-export { ReadonlyGlobalErrorCollector, createGlobalErrorCollector } from './errors';
 export { default as interpretSamlangModule } from './interpreter/source-level-interpreter';
 export { prettyPrintSamlangModule, prettyPrintHighIRModuleAsJS } from './printer';
 export { checkSources, lowerSourcesToAssemblyPrograms } from './services/source-processor';
