+++ samlang-core/package.json
@@ -3,8 +3,7 @@
   "version": "0.0.1",
   "license": "AGPLv3",
   "scripts": {
-    "compile": "tsc --incremental --noEmit",
-    "codegen": "yarn antlr4ts -o parser/generated -visitor -no-listener PL.g4 && node suppress-type-errors.js"
+    "compile": "tsc --incremental --noEmit"
   },
   "devDependencies": {
     "@types/jest": "^26.0.10",
@@ -14,6 +13,8 @@
   "dependencies": {
     "antlr4ts": "0.5.0-alpha.3",
     "samlang-core-ast": "workspace:0.0.1",
+    "samlang-core-errors": "workspace:0.0.1",
+    "samlang-core-parser": "workspace:0.0.1",
     "samlang-core-utils": "workspace:0.0.1"
   }
 }
